refactor: move db entities into the src/entites directory
