+++ src/entities/user/app.module.ts
@@ -3,8 +3,8 @@ import { AppController } from '~/app.controller'
 import { AppService } from '~/app.service'
 import { UserModule } from './user.module'
 import { PrismaModule } from '~/prisma/prisma.module'
-import { SiteModule } from '~/site/site.module'
-import { AddressModule } from '~/address/address.module'
+import { SiteModule } from '~/entities/site/site.module'
+import { AddressModule } from '~/entities/address/address.module'
 
 @Module({
   imports: [UserModule, PrismaModule, SiteModule, AddressModule],
